refactor store creation
creates store via factory function;
extracts default state and types to
separate files;
updates components accordingly;
