+++ src/store/defaultState.ts
@@ -1,0 +1,1 @@
+export const state = { error: null };
