+++ src/store/types.ts
@@ -1,4 +1,11 @@
-import { StoreState } from ".";
+import { state } from "./defaultState";
+import { ExerciseState } from "./modules/exercise/types";
+
+export type RootState = typeof state;
+
+export type StoreState = RootState & {
+  exercise: ExerciseState;
+};
 
 export type ModuleStateProps<Key extends keyof StoreState> = Readonly<
   Array<keyof StoreState[Key]>
