+++ src/views/Preferences.vue
@@ -70,7 +70,7 @@
         {{ $t("preferences.disable_start_tips") }}
       </app-checkbox>
     </div>
-	
+
     <div>
       <app-button @click="restoreDefault">
         {{ $t("preferences.restore_default") }}
@@ -83,13 +83,14 @@
 import ButtonVue from "@/components/ui/Button.vue";
 import CheckboxVue from "@/components/ui/Checkbox.vue";
 import RangeVue from "@/components/ui/Range.vue";
-import { namespaceName, StoreState } from "@/store";
+import { namespaceName } from "@/store/createStore";
 import { customizableExerciseStateProps } from "@/store/modules/exercise";
 import {
   ExerciseActions,
   ExerciseModuleMap,
   UpdatePreferencesPayload,
 } from "@/store/modules/exercise/types";
+import { StoreState } from "@/store/types";
 import { defineComponent } from "vue";
 import { mapState } from "vuex";
